Exucute mutation on submit
